#30 Change menu structure
