+++ frontend/src/components/AppMenu.js
@@ -1,4 +1,4 @@
-import { ListItemButton, ListItemIcon, ListItemText, Toolbar } from '@mui/material';
+import { IconButton, ListItemButton, ListItemIcon, ListItemText, Toolbar } from '@mui/material';
 import { Link } from "react-router-dom";
 import React from 'react';
 import CloudUploadIcon from '@mui/icons-material/CloudUpload';
@@ -8,64 +8,40 @@ import PlaylistAddCheckIcon from '@mui/icons-material/PlaylistAddCheck';
 import PlayCircleOutlineIcon from '@mui/icons-material/PlayCircleOutline';
 import SettingsIcon from '@mui/icons-material/Settings';
 import EqualizerIcon from '@mui/icons-material/Equalizer';
-function AppMenu() {
+import LayersIcon from '@material-ui/icons/Layers';
+import { getTargetIP } from '../api/API';
+import { getStatusOfIp } from '../api/data';
+
+function AppMenu({ setFalse, setTrue }) {
+    const [status, setStatus] = React.useState();
+    const ip = getTargetIP()
+    React.useEffect(() => {
+        async function fetchData() {
+            const response = await getStatusOfIp(ip)
+            return setStatus(response.data)
+        }
+        fetchData()
+    }, [ip])
+
     return (
-        <>
-            <ListItemButton component={Link} to="/">
-                <ListItemIcon>
-                    <DashboardIcon />
-                </ListItemIcon>
-                <ListItemText primary="Home" />
-            </ListItemButton>
-            <ListItemButton component={Link} to="/upload">
-                <ListItemIcon>
-                    <CloudUploadIcon />
-                </ListItemIcon>
-                <ListItemText primary="Upload" />
-            </ListItemButton>
-            <ListItemButton component={Link} to="/data">
-                <ListItemIcon>
-                    <StorageIcon />
-                </ListItemIcon>
-                <ListItemText primary="Data" />
-            </ListItemButton>
-
-            <ListItemButton component={Link} to="/status">
-                <ListItemIcon>
-                    <PlaylistAddCheckIcon />
-                </ListItemIcon>
-                <ListItemText primary="Status" />
-            </ListItemButton>
-
-            <ListItemButton component={Link} to="/predict">
-                <ListItemIcon>
-                    <PlayCircleOutlineIcon />
-                </ListItemIcon>
-                <ListItemText primary="Predict" />
-            </ListItemButton>
-
-            <ListItemButton button component={Link} to="/statics">
-                <ListItemIcon>
-                    <EqualizerIcon />
-                </ListItemIcon>
-                <ListItemText primary="Statics" />
-            </ListItemButton>
-
-            <ListItemButton component={Link} to="/settings">
-                <ListItemIcon>
-                    <SettingsIcon />
-                </ListItemIcon>
-                <ListItemText primary="Settings" />
-            </ListItemButton>
-
-
-            <ListItemButton component={Link} to="/registry-settings">
-                <ListItemIcon>
-                    <SettingsIcon />
-                </ListItemIcon>
-                <ListItemText primary="Registry Settings" />
-            </ListItemButton>
-        </>
+        <div>
+            <div onClick={setFalse}>
+                <ListItemButton component={Link} to="/">
+                    <ListItemIcon>
+                        <DashboardIcon />
+                    </ListItemIcon>
+                    <ListItemText primary="Home" />
+                </ListItemButton>
+            </div>
+            <div onClick={setTrue}>
+                <ListItemButton component={Link} to="/status">
+                    <ListItemIcon>
+                        <LayersIcon />
+                    </ListItemIcon>
+                    <ListItemText primary={`${status?.name}`} />
+                </ListItemButton>
+            </div>
+        </div>
     )
 };
