+++ frontend/src/screens/StatusScreen.js
@@ -54,13 +54,13 @@ const StatusScreen = () => {
     return (
         <>
             {status &&
-                <Grid container alignItems="center" className={classes.grid} direction="column">
+                <Grid container alignItems="flex-start" className={classes.grid} direction="column">
                     <Grid item container className={classes.gridRow} >
-                        <Typography style={{ marginLeft: '-130px' }}>Status:</Typography>
+                        <Typography style={{ marginLeft: '-530px' }}>Status:</Typography>
                         <Typography>{status?.state_name}</Typography>
                     </Grid>
                     <Grid item container className={classes.gridRow}>
-                        <Typography style={{ marginLeft: '-70px' }}>Format:</Typography>
+                        <Typography style={{ marginLeft: '-470px' }}>Format:</Typography>
                         <Typography>{format}</Typography>
                     </Grid>
                     <Button variant="outlined" className={classes.button} onClick={handleClick}>Rebuild</Button>
